refactor: move theme into bccm_player export
