+++ lib/bccm_player.dart
@@ -25,5 +25,9 @@ export 'src/widgets/controls/play_next_button.dart';
 export 'src/widgets/utils/bccm_player_state_builder.dart';
 export 'src/widgets/controls/player_controls.dart';
 export 'src/model/player_view_config.dart';
+export 'src/theme/controls_theme_data.dart';
+export 'src/theme/bccm_player_theme.dart';
+export 'src/theme/player_theme.dart';
+export 'src/theme/mini_player_theme_data.dart';
 export 'src/pigeon/pigeon_extensions.dart';
 export 'src/utils/time.dart' show calcTimeLeftMs;
