+++ lib/src/theme/player_theme.dart
@@ -1,6 +1,6 @@
 import 'package:bccm_player/src/utils/extensions.dart';
-import 'package:bccm_player/theme/controls_theme_data.dart';
-import 'package:bccm_player/theme/mini_player_theme_data.dart';
+import 'package:bccm_player/src/theme/controls_theme_data.dart';
+import 'package:bccm_player/src/theme/mini_player_theme_data.dart';
 import 'package:flutter/material.dart';
 
 class PlayerTheme extends InheritedWidget {
