+++ lib/src/widgets/mini_player/mini_player.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../../../theme/player_theme.dart';
+import '../../theme/player_theme.dart';
 import '../../utils/svg_icons.dart';
 import '../../utils/transparent_image.dart';
 import 'loading_indicator.dart';
